重构t_config表

+++ cat-core/src/main/java/com/dianping/cat/config/ReportReloadConfigManager.java
@@ -50,7 +50,7 @@ public class ReportReloadConfigManager {
 
 	protected ContentFetcher m_fetcher;
 
-	private int m_configId;
+	private long m_configId;
 
 	private long m_modifyTime;
