+++ cat-home/src/main/java/com/dianping/cat/report/page/dependency/config/TopoGraphFormatConfigManager.java
@@ -42,7 +42,7 @@ public class TopoGraphFormatConfigManager {
 
 	private ContentFetcher m_fetcher;
 
-	private int m_configId;
+	private long m_configId;
 
 	private TopoGraphFormatConfig m_config;
